--- app/src/main/java/com/project/lumina/client/game/module/effect/LevitationElement.kt
@@ -1,12 +1,0 @@
-package com.project.lumina.client.game.module.effect
-
-import com.project.lumina.client.R
-import com.project.lumina.client.game.data.Effect
-import com.project.lumina.client.game.module.setting.EFFECTS
-
-class LevitationElement : BaseEffectElement(
-    name = "levitation",
-    displayNameResId = R.string.module_levitation_display_name,
-    effectId = Effect.LEVITATION,
-    effectSetting = EFFECTS.LEVITATION
-)
